pyproject.toml
[tool.poetry]
name = "autosnab-bot"
version = "0.1.0"
description = ""
authors = ["Kasyanov Vladimir <podriv-ustoev@yandex.ru>"]
readme = "README.md"
packages = [{include = "autosnab_bot"}]

[tool.poetry.dependencies]
python = "^3.11"

python-telegram-bot = "==20.0"
python-dotenv = "==0.21.1"
python-calamine = "==0.3.1"
prettytable = "==3.11.0"

[tool.poetry.group.dev.dependencies]
ruff = "==0.0.240"
pyright = "==1.1.389"
black = "==23.1.0"
pytest = "==8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["F", "E", "W", "C90",
          "I", "N", "S", "B", "A",
          "ISC", "T20", "Q", "PTH"]
line-length = 120

[tool.ruff.per-file-ignores]
"tests/*.py" = ["S101"]

ignore = ["A003"]

[tool.pyright]
reportUnnecessaryTypeIgnoreComment="warning"
